+++ src/Infrastructure/MessageSerialization/Exceptions/EncodeMessageFailed.php
@@ -11,14 +11,14 @@
 
 declare(strict_types = 1);
 
-namespace Desperado\ServiceBus\Marshal\Serializer\Exceptions;
+namespace Desperado\ServiceBus\Infrastructure\MessageSerialization\Exceptions;
 
 use Desperado\ServiceBus\Common\Exceptions\ServiceBusExceptionMarker;
 
 /**
  *
  */
-final class SerializationFailed extends \RuntimeException implements ServiceBusExceptionMarker
+final class EncodeMessageFailed extends \RuntimeException implements ServiceBusExceptionMarker
 {
 
 }