+++ src/Transport/IncomingEnvelope.php
@@ -28,13 +28,6 @@ final class IncomingEnvelope
     private $plain;
 
     /**
-     * Normalized message representation
-     *
-     * @var array
-     */
-    private $normalized;
-
-    /**
      * Denormalized message representation
      *
      * @var Message
@@ -51,14 +44,12 @@ final class IncomingEnvelope
     /**
      * @param string  $operationId
      * @param string  $plain
-     * @param array   $normalized
      * @param Message $decoded
      * @param array   $headers
      */
-    public function __construct(string $plain, array $normalized, Message $decoded, array $headers = [])
+    public function __construct(string $plain, Message $decoded, array $headers = [])
     {
         $this->plain       = $plain;
-        $this->normalized  = $normalized;
         $this->decoded     = $decoded;
         $this->headers     = $headers;
     }
@@ -71,16 +62,6 @@ final class IncomingEnvelope
     public function requestBody(): string
     {
         return $this->plain;
-    }
-
-    /**
-     * Receive normalized message representation
-     *
-     * @return array
-     */
-    public function normalized(): array
-    {
-        return $this->normalized;
     }
 
     /**